requirements.txt
Flask==2.2.3
flask-cors==3.0.10
flask-limiter==3.3.1
openai==0.27.8
pydantic==1.10.8
python-dotenv==1.0.0
marshmallow==3.19.0
gunicorn==20.1.0
pytest==7.3.1
pytest-cov==4.1.0 